Added in commands that set the gains for spin and scan mode.
Also, changed some of the gain terms for spin and scan mode.  Added a
term that will be proportional to the gondola velocity derivative for
the reaction wheel, (still need to write the code that calculates the
velocity derivative).

Wrote the gains to the frame in slowMotorFields in motors.cpp, the
scaling factors/limits were stored in motordefs.h