Tested enable/disable functionality for the Copley controllers.
reactComm now monitors the velocity and encoder position of the reaction wheel
if the reaction wheel configuration was sucessful.

Wrote reaction wheel velocity and position to the frame and added an
extra loop node (LOOP7).  Still need to test the RW vel & position
written to the frame to make sure it is correct.

Tested without power to the reaction wheel controller.  mcp complains but
still runs fine.