Added in error handling for the motor serial communication threads.  Only fully implemented for the Reaction Wheel motor at present.
Added a check to the motor threads that waits until the computer running
mcp is in charge before attempting to communicate with the motors.

Also added in a couple of extra sleep calls and minimized print
statements so that dgps.c doesn't competely fill the buffer when the
serial hub is not powered.