Adding robust protection to gyro readout.
It was observed that multiple programs could access the gyro stream (by
intent) but that if this occurred while MCP was running that the system
would cease reading valid data from the gyro stream.  This can occur
when another thread requests exclusive access or injects garbage into
the stream.

We resolve this by disconnecting and reconnecting the stream when we
receive more than 2 packets worth of garbage.  We also provide this
function when the system times out (1 second currently) or receives
errors in the read.