Update watchdog tickle to fix crash and bi0 drop (some)
The watchdog tickle acts as a cached read.  Therefore, we can't pass a
stack pointer as the output memory location (needs to be static).  We
also need to keep the output state of the TDI/TCK pins as failing to do
that will mess up the bi0.  

The solution is to read the current state of the pins, toggle the tickle
pin, then write the new state back.  The pin state is stored in a static
variable.